fix: resolve backend-size violations for cursor session metrics PR
Split cursorstorage store loading and parsing across focused files,
extract Cursor inference completion from Execute, and trim provider
tests so make backend-size passes on the reviewed head.
